Upgrade the Monaco Editor to version 0.9.0
The facade is generated with
sjrd/scala-js-ts-importer#a2a2ce740d66adeb9fb53b7be7799dfe402febab which
uses `def` instead of `val` for readonly fields compared to the facade
generated for 0.8.3.

+++ metadoc-js/src/main/scala/metadoc/ScalaReferenceProvider.scala
@@ -20,7 +20,7 @@ class ScalaReferenceProvider(index: Index) extends ReferenceProvider {
       context: ReferenceContext,
       token: CancellationToken
   ) = {
-    val offset = model.getOffsetAt(position)
+    val offset = model.getOffsetAt(position).toInt
     for {
       attrs <- MetadocAttributeService.fetchAttributes(model.uri.path)
       id = IndexLookup.findSymbol(offset, attrs, index).map(_.symbol)
